backend/requirements.txt
PyYAML>=6.0

Pillow>=9.0.0

zhipuai>=2.0.0

python-dotenv>=1.0.0

gradio

PyQt5>=5.15.0

opencv-python>=4.8.0